batch/pyproject.toml
[project]
name = "batch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
    "tenacity>=8.2.0",
    "python-dateutil>=2.9.0.post0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "SQLAlchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.1",
]
